implement test_tslib, move test_normalize_date from test_tools
